+++ packages/fence_service/lib/weebi_app_service.dart
@@ -1,3 +1,1 @@
-library;
-
 export 'src/weebi_app_service/weebi_app_service_base.dart';